Rename lib to src
