+++ .flowconfig
@@ -8,7 +8,7 @@
 <PROJECT_ROOT>/bin
 
 [include]
-<PROJECT_ROOT>/lib
+<PROJECT_ROOT>/src
 .*/__tests__/.*
 
 [options]
