+++ bin/checksync.dev.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 require("@babel/register");
-require("../lib/cli.js").run(__filename);
+require("../src/cli.js").run(__filename);
