+++ jest.config.js
@@ -14,7 +14,7 @@ module.exports = {
     // Setting this value to "fake" allows the use of fake timers for functions such as "setTimeout"
     timers: "fake",
 
-    setupFiles: ["./lib/shims.js"],
+    setupFiles: ["./src/shims.js"],
 
     setupFilesAfterEnv: ["jest-extended"],
 };
