+++ src/__tests__/checksum_test.js
