+++ src/__tests__/cli_test.js
