+++ src/__tests__/get-markers-from-files_test.js
