+++ src/check-sync.js
