+++ src/checksum.js
