+++ src/get-markers-from-files.js
