+++ src/parse-file.js
