+++ src/scoped-logger.js
